refactor: Lazy loading the unneeded components
