+++ assets/controllers.json
@@ -3,7 +3,7 @@
     "@symfony/ux-chartjs": {
       "chart": {
         "enabled": true,
-        "fetch": "eager"
+        "fetch": "lazy"
       }
     },
     "@symfony/ux-live-component": {
