+++ assets/controllers/challenge_instruction_modal_controller.ts
@@ -1,17 +1,18 @@
 import { Controller } from "@hotwired/stimulus";
 import { Component, getComponent } from "@symfony/ux-live-component";
-import * as bootstrap from "bootstrap";
+import type * as bootstrap from "bootstrap";
 
 export default class extends Controller<HTMLElement> {
   #component: Component | undefined;
   #modal: bootstrap.Modal | undefined;
 
-  async initialize(): Promise<void> {
+  async initialize() {
     this.#component = await getComponent(this.element);
   }
 
-  connect(): void {
-    this.#modal = new bootstrap.Modal(this.element);
+  async connect() {
+    const bs = await import("bootstrap");
+    this.#modal = new bs.Modal(this.element);
   }
 
   async open() {
