+++ assets/styles/bootstrap.scss
@@ -1,0 +1,61 @@
+@use "sass:list";
+@import "../../vendor/twbs/bootstrap/scss/functions";
+
+// Variables
+$prefix: bs-;
+$spacer: 0.8rem;
+$primary: #4154f1;
+$navbar-padding-x: 2rem;
+$navbar-padding-y: $spacer * 0.85;
+$input-btn-border-width: var(--#{$prefix}border-width);
+$input-border-width: $input-btn-border-width;
+$input-height-border: calc(#{$input-border-width} * 2);
+$form-floating-height: add(3.5rem * 0.8, $input-height-border);
+$form-floating-padding-y: 0.65rem;
+
+$timeline-color: #808080;
+$timeline-width: 2px;
+
+@import "../../vendor/twbs/bootstrap/scss/variables";
+
+$font-family-sans-serif: list.join(
+  ("LXGW WenKai TC", cursive),
+  $font-family-sans-serif
+);
+
+// required parts
+@import "../../vendor/twbs/bootstrap/scss/maps";
+@import "../../vendor/twbs/bootstrap/scss/mixins";
+@import "../../vendor/twbs/bootstrap/scss/root";
+@import "../../vendor/twbs/bootstrap/scss/utilities";
+
+// other components
+@import "../../vendor/twbs/bootstrap/scss/reboot";
+@import "../../vendor/twbs/bootstrap/scss/type";
+@import "../../vendor/twbs/bootstrap/scss/images";
+@import "../../vendor/twbs/bootstrap/scss/containers";
+@import "../../vendor/twbs/bootstrap/scss/grid";
+@import "../../vendor/twbs/bootstrap/scss/tables";
+@import "../../vendor/twbs/bootstrap/scss/forms";
+@import "../../vendor/twbs/bootstrap/scss/buttons";
+@import "../../vendor/twbs/bootstrap/scss/transitions";
+@import "../../vendor/twbs/bootstrap/scss/dropdown";
+@import "../../vendor/twbs/bootstrap/scss/button-group";
+@import "../../vendor/twbs/bootstrap/scss/nav";
+@import "../../vendor/twbs/bootstrap/scss/navbar";
+@import "../../vendor/twbs/bootstrap/scss/card";
+@import "../../vendor/twbs/bootstrap/scss/breadcrumb";
+@import "../../vendor/twbs/bootstrap/scss/pagination";
+@import "../../vendor/twbs/bootstrap/scss/badge";
+@import "../../vendor/twbs/bootstrap/scss/alert";
+@import "../../vendor/twbs/bootstrap/scss/list-group";
+@import "../../vendor/twbs/bootstrap/scss/close";
+@import "../../vendor/twbs/bootstrap/scss/modal";
+@import "../../vendor/twbs/bootstrap/scss/tooltip";
+@import "../../vendor/twbs/bootstrap/scss/spinners";
+
+// Helpers
+@import "../../vendor/twbs/bootstrap/scss/helpers";
+
+// Utilities
+@import "../../vendor/twbs/bootstrap/scss/utilities/api";
